+++ src/main/java/com/anoop/rl/repository/UserRepository.java
@@ -11,9 +11,9 @@ import com.anoop.rl.model.UserEntity;
 
 @Repository
 @Transactional
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
-    Optional<UserEntity> findById(Integer id);
+    Optional<UserEntity> findById(Long id);
     Optional<UserEntity> findByUsername(String username);
     Optional<UserEntity> findByEmail(String email);
     boolean existsByUsername(String username);
